+++ lib/mutant/mutator/node/define.rb
@@ -14,7 +14,6 @@ module Mutant
         def dispatch
           emit_arguments_mutations
           emit_optarg_body_assignments
-          emit_restarg_body_mutation
           emit_body(N_RAISE)
           emit_body(N_ZSUPER)
           emit_body(nil)
@@ -29,17 +28,6 @@ module Mutant
             next unless n_optarg?(argument) && AST::Meta::Optarg.new(argument).used?
 
             emit_body_prepend(s(:lvasgn, *argument))
-          end
-        end
-
-        # Emit mutation with arg splat as empty array assignment in method
-        #
-        # @return [undefined]
-        def emit_restarg_body_mutation
-          arguments.children.each do |argument|
-            next unless n_restarg?(argument) && argument.children.one?
-
-            emit_body_prepend(s(:lvasgn, AST::Meta::Restarg.new(argument).name, s(:array)))
           end
         end