+++ meta/def.rb
@@ -8,31 +8,6 @@ Mutant::Meta::Example.add :def do
 end
 
 Mutant::Meta::Example.add :def do
-  source 'def foo(a, *b); nil; end'
-
-  mutation 'def foo(_a, *b); nil; end'
-  mutation 'def foo; nil; end'
-  mutation 'def foo(a, *b); end'
-  mutation 'def foo(a, *b); raise; end'
-  mutation 'def foo(a); nil; end'
-  mutation 'def foo(*b); nil; end'
-  mutation 'def foo(a, *b); b = []; nil; end'
-  mutation 'def foo(a, *b); super; end'
-end
-
-Mutant::Meta::Example.add :def do
-  source 'def foo(a, *); nil; end'
-
-  mutation 'def foo(_a, *); nil; end'
-  mutation 'def foo; nil; end'
-  mutation 'def foo(a, *); end'
-  mutation 'def foo(a, *); raise; end'
-  mutation 'def foo(a); nil; end'
-  mutation 'def foo(*); nil; end'
-  mutation 'def foo(a, *); super; end'
-end
-
-Mutant::Meta::Example.add :def do
   source 'def foo; foo; rescue; end'
 
   mutation 'def foo; raise; end'